Mudou coisa bagarai sla man

+++ petcemetery/src/main/java/com/petcemetery/petcemetery/repositorio/ServicoRepository.java
@@ -8,7 +8,7 @@ import com.petcemetery.petcemetery.model.Servico;
 
 
 public interface ServicoRepository extends JpaRepository<Servico, Long> {
-    Servico findByIdServico(int idServico);
+    Servico findByIdServico(Long idServico);
     List<Servico> findBycliente_cpf(String cpf);
 
 }
